+++ lib/formats.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Id: git-versioning/0.0.31-dev+20160321-0713 lib/formats.sh
+# Id: git-versioning/0.0.31-dev+20160412-1532 lib/formats.sh
 
 # reStructureText
 RST_VER_TOKEN=':\([Vv]\)ersion:'