+++ tools/cmd/src-files.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Id: git-versioning/0.0.31-dev+20160321-0713 tools/cmd/src-files.sh
+# Id: git-versioning/0.0.31-dev+20160412-1532 tools/cmd/src-files.sh
 
 # Seed code for 'src-files' command. Lists (versioned) src files.