+++ tools/git-hooks/pre-commit.sh
@@ -1,2 +1,2 @@
-# Id: git-versioning/0.0.31-dev+20160321-0713 tools/git-hooks/pre-commit.sh
+# Id: git-versioning/0.0.31-dev+20160412-1532 tools/git-hooks/pre-commit.sh
 make git-pre-commit