+++ ChangeLog.rst
@@ -20,4 +20,9 @@ Change Log
   - Documentation restructured. Fixes, updates.
     Added reference to project GitVersion.
   - Removed V_PATH_LIST. Started read_nix_style_file.
+  - Added modus without versioned-files.list (and main document with version).
+    VER_STR must be provided in the environment, the files to be checked are
+    read from standard-input::
+
+      VER_STR=1.2.3-alpha git-versioning check